add additional booleans
