+++ lib/attribute_display.dart
@@ -1,12 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
-class AttributeOption {
-  final String key;
-  final IconData icon;
+import 'attribute_data.dart'; // Import the shared data
 
-  const AttributeOption(this.key, this.icon);
-}
 
 class AttributeDisplay extends StatelessWidget {
   final List<String> details;
@@ -22,48 +17,17 @@ class AttributeDisplay extends StatelessWidget {
     this.onEditTap,
   }) : super(key: key);
 
-  static const List<AttributeOption> detailOptions = [
-    AttributeOption('indoor', FontAwesomeIcons.house),
-    AttributeOption('accessible', FontAwesomeIcons.wheelchair),
-    AttributeOption('trees', FontAwesomeIcons.tree),
-    AttributeOption('grass', FontAwesomeIcons.pagelines),
-  ];
-
-  static const List<AttributeOption> amenityOptions = [
-    AttributeOption('restroom', FontAwesomeIcons.toilet),
-    AttributeOption('drinking_fountain', FontAwesomeIcons.faucet),
-    AttributeOption('dog_park', FontAwesomeIcons.dog),
-    AttributeOption('seating', FontAwesomeIcons.chair),
-    AttributeOption('tables', FontAwesomeIcons.table),
-    AttributeOption('wifi', FontAwesomeIcons.wifi),
-    AttributeOption('art', FontAwesomeIcons.paintBrush),
-    AttributeOption('parking', FontAwesomeIcons.squareParking),
-    AttributeOption('picnic_shelter', FontAwesomeIcons.umbrellaBeach),
-    AttributeOption('chess_tables', FontAwesomeIcons.chessBoard),
-    AttributeOption('fountain', FontAwesomeIcons.water),
-    AttributeOption('food_vendor', FontAwesomeIcons.hotdog),
-  ];
-
-  static const List<AttributeOption> equipmentOptions = [
-    AttributeOption('playground', FontAwesomeIcons.children),
-    AttributeOption('splash_pad', FontAwesomeIcons.water),
-    AttributeOption('exercise', FontAwesomeIcons.dumbbell),
-    AttributeOption('handball', FontAwesomeIcons.baseball),
-    AttributeOption('basketball', FontAwesomeIcons.basketball),
-    AttributeOption('baseball', FontAwesomeIcons.baseballBatBall),
-    AttributeOption('track', FontAwesomeIcons.personRunning),
-  ];
 
   List<Widget> _buildChips(
       List<String> activeKeys, List<AttributeOption> options) {
     return options
         .where((option) => activeKeys.contains(option.key))
         .map((option) => Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 4, vertical: 4),
+              padding: const EdgeInsets.symmetric(horizontal: 4, vertical: 0),
               child: Chip(
-                avatar: Icon(option.icon, size: 14, color: Colors.black),
+                avatar: Icon(option.icon, size: 14, color: Colors.grey[700]),
                 label: Text(option.key.replaceAll('_', ' ').toUpperCase()),
-                labelStyle: const TextStyle(fontSize: 11, color: Colors.black),
+                labelStyle: TextStyle(fontSize: 11, color: Colors.grey[700]),
                 backgroundColor: Colors.white,
                 side: BorderSide.none,
               ),
@@ -84,7 +48,6 @@ class AttributeDisplay extends StatelessWidget {
         crossAxisAlignment: CrossAxisAlignment.stretch,
         children: [
           Text(title, style: const TextStyle(fontWeight: FontWeight.bold)),
-          const SizedBox(height: 8),
           if (items.isNotEmpty)
             Wrap(
               alignment: WrapAlignment.start,
@@ -101,8 +64,8 @@ class AttributeDisplay extends StatelessWidget {
                   visualDensity: VisualDensity.compact,
                 ),
                 onPressed: onEditTap,
-                icon: const Icon(Icons.add, size: 16),
-                label: Text(addLabel, style: const TextStyle(fontSize: 14)),
+                icon:  Icon(Icons.add, size: 16, color: Colors.grey[700]),
+                label: Text(addLabel, style: TextStyle(fontSize: 14, color: Colors.grey[700])),
               ),
             ),
         ],
@@ -118,19 +81,19 @@ class AttributeDisplay extends StatelessWidget {
         _buildSection(
           title: 'Details',
           items: details,
-          options: detailOptions,
+          options: AttributeData.detailOptions,
           addLabel: 'Add details',
         ),
         _buildSection(
           title: 'Amenities',
           items: amenities,
-          options: amenityOptions,
+          options: AttributeData.amenityOptions,
           addLabel: 'Add amenities',
         ),
         _buildSection(
           title: 'Equipment',
           items: equipment,
-          options: equipmentOptions,
+          options: AttributeData.equipmentOptions,
           addLabel: 'Add equipment',
         ),
       ],
